requirements.txt
msal==1.31.0
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.5
python-pptx==0.6.23
pillow==10.4.0
boto3==1.35.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
